Moving manifest into subnamespace
