+++ src/mantle/application/class-application.php
@@ -12,8 +12,8 @@ use Mantle\Contracts\Application as Application_Contract;
 use Mantle\Contracts\Container as Container_Contract;
 use Mantle\Contracts\Kernel as Kernel_Contract;
 use Mantle\Contracts\Support\Isolated_Service_Provider;
-use Mantle\Framework\Model_Manifest;
-use Mantle\Framework\Package_Manifest;
+use Mantle\Framework\Manifest\Model_Manifest;
+use Mantle\Framework\Manifest\Package_Manifest;
 use Mantle\Log\Log_Service_Provider;
 use Mantle\Events\Event_Service_Provider;
 use Mantle\Framework\Providers\Routing_Service_Provider;
