+++ src/mantle/framework/providers/class-model-service-provider.php
@@ -9,7 +9,7 @@ namespace Mantle\Framework\Providers;
 
 use Mantle\Database\Model\Model;
 use Mantle\Database\Model\Relations\Relation;
-use Mantle\Framework\Model_Manifest;
+use Mantle\Framework\Manifest\Model_Manifest;
 use Mantle\Support\Service_Provider;
 
 /**
